chore(axeptio): move sdk loader to separate file
