+++ src/integrations/Axeptio/browser.js
@@ -2,8 +2,8 @@
 /* eslint-disable class-methods-use-this */
 import { NAME } from './constants';
 import Logger from '../../utils/logger';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import makeACall from './utils';
+import { loader } from './loader';
 
 const logger = new Logger(NAME);
 
@@ -21,22 +21,9 @@ class Axeptio {
     this.destinationId = destinationInfo && destinationInfo.destinationId;
   }
 
-  loadScript() {
-    window.axeptioSettings = {
-      clientId: this.clientId,
-    };
-    (function (d, s) {
-      var t = d.getElementsByTagName(s)[0],
-        e = d.createElement(s);
-      e.async = true;
-      e.src = '//static.axept.io/sdk.js';
-      e.setAttribute('data-loader', LOAD_ORIGIN), t.parentNode.insertBefore(e, t);
-    })(document, 'script');
-  }
-
   init() {
     logger.debug('===In init Axeptio===');
-    this.loadScript();
+    loader(document, 'script', this.clientId);
   }
 
   isLoaded() {
